Move effective retention tests to BasicDataStreamLifecycle
